added the 'preventing replay' section for alex
Our protocol (if you implement it as i described it, with the last step being AES not RSA) provides all of the required components and with minor modification will provide True Authenticity (i.e. prevent man-in-the-middle)

We can provide true authenticity by making sure eveyone only ever uses one RSA key (the same key to connect to server and to handshake with clients), the server is in charge of relaying public keys to clients ON THEIR BEHALF when a client-client communication is requested and then the clients proceed skipping the initial sending of public keys in the clear.

This effectively says, if the server is given limited trust, then all clients who i initiate communication with over the server are completely trusted

But since this level of authenticity is not required (we only have to ensure public key exchange does not compromise security) this is an extension, shouldnt waste time on it